refactor + blog added
