+++ src/routes/(app)/+layout.svelte
