+++ src/routes/(app)/blog/+layout@.svelte
@@ -1,0 +1,6 @@
+<script>
+	import Navbar from '$lib/components/Layout/Navbar.svelte';
+</script>
+
+<Navbar />
+<slot />
