+++ src/routes/(app)/blog/[slug]/+page.server.js
@@ -1,0 +1,15 @@
+import { getBlogPost } from '$lib/backend-requests.js';
+
+export async function load({ params }) {
+	let post = null;
+
+	try {
+		post = await getBlogPost(params.slug);
+	} catch (e) {
+		console.error(e);
+	}
+
+	return {
+		post: post
+	};
+}
